+++ demos/chinook/src/main/java/is/codion/demos/chinook/model/TrackTableModel.java
@@ -23,7 +23,7 @@ import is.codion.demos.chinook.domain.api.Chinook.Track.RaisePriceParameters;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
-import is.codion.framework.model.EntityTableConditionModel;
+import is.codion.framework.model.EntityConditionModel;
 import is.codion.framework.model.ForeignKeyConditionModel;
 import is.codion.swing.framework.model.SwingEntityConditionModelFactory;
 import is.codion.swing.framework.model.SwingEntityTableModel;
@@ -42,7 +42,7 @@ public final class TrackTableModel extends SwingEntityTableModel {
 
 	public TrackTableModel(EntityConnectionProvider connectionProvider) {
 		super(new TrackEditModel(connectionProvider),
-						entityQueryModel(EntityTableConditionModel.builder()
+						entityQueryModel(EntityConditionModel.builder()
 										.entityType(Track.TYPE)
 										.connectionProvider(connectionProvider)
 										.conditionModelFactory(new TrackColumnConditionFactory(connectionProvider))